feat(ZH-96): 효율적인 apis 파일 관리를 위한 index 파일
