+++ src/apis/index.ts
@@ -1,11 +1,1 @@
-import axios from 'axios';
-
-const instance = axios.create({
-  baseURL: 'http://localhost:8080',
-  headers: {
-    'Content-Type': 'application/json',
-  },
-  withCredentials: true,
-});
-
-export default instance;
+export * from './like';
