+++ src/apis/like.ts
@@ -1,0 +1,5 @@
+import instance from './instance';
+
+export const toggleLike = (productId) => {
+  return instance.get(`/products/${productId}/like`);
+};
